moving isTestAccountPhone from app to domain
